add new ReadPlaylistFragment / rename other library fragments
