+++ app/src/main/java/org/guru/playlistmaker/ui/library/libraryPlaylist/playlistAdapter/PlaylistViewHolder.kt
@@ -1,8 +1,7 @@
-package org.guru.playlistmaker.ui.library.playlist.playlistAdapter
+package org.guru.playlistmaker.ui.library.libraryPlaylist.playlistAdapter
 
 import android.content.Context
 import android.net.Uri
-import android.util.Log
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
