+++ app/src/main/java/org/guru/playlistmaker/ui/library/readPlaylist/view_model/ReadPlaylistViewModel.kt
@@ -1,0 +1,7 @@
+package org.guru.playlistmaker.ui.library.readPlaylist.view_model
+
+import androidx.lifecycle.ViewModel
+import org.koin.core.component.KoinComponent
+
+class ReadPlaylistViewModel : ViewModel(), KoinComponent {
+}
