+++ app/src/main/res/navigation/playlist_maker_navigation.xml
@@ -18,7 +18,7 @@
         android:id="@+id/mediaLibraryFragment"
         android:name="org.guru.playlistmaker.ui.library.fragment.MediaLibraryFragment"
         android:label="fragment_media_library"
-        tools:layout="@layout/fragment_media_library" >
+        tools:layout="@layout/fragment_library" >
         <action
             android:id="@+id/action_mediaLibraryFragment_to_playerFragment"
             app:destination="@id/playerFragment" />
